Basic validation for New Event Type to prevent unexpected behaviour
Previously, when creating a new event type, user was still allowed to create a new event type, even if the event length's field was empty. This would cause a prisma error, since it would receive NaN and no event would have been added, but the page still reloaded. Therefore, a very basic validation to address this, at least in the short-term.